update project, replace Controller with RestController and update URL structure

+++ scr/main/java/app/api/entity/UserId.java
@@ -1,4 +1,3 @@
 package app.api.entity;
 
-public record UserId(int id) {
-}
+public record UserId(long id) {}
